OF-857: Added offline packet deliverer
When a connection is closed, any further packets to be delivered need to
be handled by the "backup" packet deliverer. The original implementation
simply tried to reuse the primary delivery/routing mechanism, which
resulted in a recursive call. This has been replaced with a packet
deliverer that will route undeliverable message packets into the offline
store.