Use real Java enums for presence subscription states.
The homemade IntEnum and Enum classes feel totally weird and confusing, when having java.lang.Enum around.

This commit removes these classes.

The static constants in the RosterItem class now feel superfluous, but are kept in there, because they are referenced by a lot of other classes and JSP pages.