OF-984 Avoid using interned domains as sync keys
A few places all use the domain.intern() as the key for a synchronized block.

This can lead to clashes when multiple semantics conflict, and this can in turn
lead to a deadlock, as witnessed in OF-984.

This changes two cases to use a prefixed string to namespace the locking.