OF-1203: Prevent plugin (un)load concurrency issues
This commit removes direct interaction with the plugin manager state, when
working with the admin console. Instead, such actions will now cause the
plugin monitor to run, which will indirectly apply the desired changes to
the state of the plugin manager. This avoid concurrency issues in situations
where the admin console and plugin monitor attempt to modify the state of
the plugin manager at the same time.