OF-546: Plugins should not inherit from parent.
The build plans for the plugins should not inherit from the parent project, as this
ties their existence in with our parent project. Third party developers cannot do
this (without modifying the parent project structure). We should facilitate third
party developers, and use the same tools as we provide them.

This commit removes the parent/child dependencies between the top-level parent POM
and the plugins POM. There still as a parent/child dependency between the plugins
POM and each individual plugin.

Although I'm not sure this change is a complete solution, it is aimed to more closely
represent a development setup in which a third party developer might find itself. To
create a plugin of their own, they can now:

 - Use our 'plugins' module directly (define it as a parent)
 - Replace our 'plugins' module by a proprietary one (when developing multiple
   proprietary plugins).
 - Take most POM definition from our 'plugins' module and merge that with a new
   stand-alone plugin project.