OF-546: Distribute custom tags in a library.
For plugins to be able to use the tags used in the admin console, the tags should be distributed in a library. By making this library a Maven module, plugins can define the library as a dependency.

Note that the admin.tld file is duplicated, but *slightly* modified in the Maven module: the path of the tags is different (META-INF vs WEB-INF).