OF-1471: Terminate upon invalid Stream Management data.
When a client tells us that it received an amount of stanzas that's greater than the number of stanzas
that we've sent it, the stream should be terminated. Similarly, we should not allow a client to resume
a stream when it tells us it received more data on the old stream than what we've sent it.