OF-1545: Users of AuthToken.getUsername() should be able to handle null values
The Javadoc of getUsername() clearly states that null values are to be expected (when anonymous
authentication took place). In a previous commit, an instance were the username was filled with
the anonymous resource-part was corrected (now, null is used instead). This commit fixes a few
instances where users of getUsername() assumed its return value to be non-null.