OF-1383: Configure java modules, when appropriate.
This commit changes the configuration of the compiler, to work with java modules.

To remain compatible with versions of Java that do not have the modules feature (everything up to Java 8), two profiles have been created. Depending on the version of Java being used, exactly one of these is activated.