OF-1590 Validate h against resumed session
The `<resume/>` element sent by the client during a XEP-0198 session
resumption contains an `h` attribute which is validated to ensure it
has not acknowledged more stanzas than were actually sent.

However, we were validating that against the current session, whose
counters would be disabled, and therefore left at zero. This simply
checks them against the correct session instead - the `otherSession`
which is to be resumed.