OF-1603: Self-signed certs should be regenerated when the existing one does not match all server identities.
The self-signed certificate that's installed in Openfire should be replaced (when the user instructs to do so)
when the existing certificate does not include all server identities as subject alternate names (of the DNS type).

The existing code checked only if the primary identity was present in the certificate, and ignored the request
to replace the certificate if it was.