Remove an orphan component cache entry before adding a new one.
The orphan component cache entry had a DEFAULT_NODE_ID as node ID
which was added before Hazelcast was initialized with a real node ID.
When the real node ID was ready and it was added to component cache
without removing the old entry.  So the component cache always had two
entries for a component.  When the component was disconnected, the
Component Manager believed that the component existed and prevented the
component to reconnect.