OF-1700: When stopping the manager, cleanly shut down its components.
When a manager of components is stopped, its managed components should also be gracefully stopped.

My assumption is that this didn't occur, as ordinarily, stopping a manager indicates a shutdown of Openfire, in which all components are stopped anyways.

In a clustered setup, it is desirable to have components shut down gracefully. An important effect of this is that other cluster nodes are notified of the shutdown.