Introduce a new API to provide locking of Cache entries.
Why? The single responsibility principle. It makes more sense (to me)
to ask the Cache to lock a key rather than the CacheFactory, which should
concentrate on creating cache's.

Note that only the API changes; the implementation still resides in
CacheFactory. Easiest to leave it there until the deprecated method is
removed.