OF-1793: Contact List Sharing for LDAP groups shows wrong setting.
Groups imported through LDAP initially have no properties. The admin console checks the Contact List Sharing configuration of a group, by evaluating the value of the sharedRoster.showInRoster property. Previously, this needed to have the value 'nobody' for the group to be marked as 'unshared'. This commit ensures that a group is also shown to be 'unshared' if the property is not set at all.